buffer: add explicit cast to remove narrowing warning (#19879)

+++ source/common/buffer/buffer_impl.h
@@ -369,7 +369,7 @@ public:
    */
   static inline SizedStorage newStorage(uint64_t min_capacity) {
     const uint64_t slice_size = sliceSize(min_capacity);
-    return {StoragePtr{new uint8_t[slice_size]}, slice_size};
+    return {StoragePtr{new uint8_t[slice_size]}, static_cast<size_t>(slice_size)};
   }
 
 protected:
